fix: move to spring-data-mongodb
