+++ persistence-modules/spring-data-mongodb/src/main/java/com/baeldung/projection/model/InStock.java
